Разобьет код на компоненты, добавит webpack
