+++ public/index.html
@@ -42,7 +42,7 @@
       </div>
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
 
   </body>
 </html>
